+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/util/Images.android.kt
@@ -9,3 +9,4 @@ actual suspend fun getImageDimensions(byteArrayFlow: ByteArrayFlow): Pair<Int?, 
     val bitmap = BitmapFactory.decodeByteArray(byteArray, 0, byteArray.size)
     return bitmap?.width to bitmap?.height
 }
+