+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/SendAttachmentViewModel.kt
@@ -4,11 +4,13 @@ import com.arkivanov.essenty.backhandler.BackCallback
 import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.util.FileDescriptor
 import de.connect2x.trixnity.messenger.util.ManualFileDescriptor
+import de.connect2x.trixnity.messenger.util.ProcessImageUpload
 import de.connect2x.trixnity.messenger.util.getImageDimensions
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
 import de.connect2x.trixnity.messenger.viewmodel.i18n
 import de.connect2x.trixnity.messenger.viewmodel.util.checkFileSizeExceedsLimit
 import io.github.oshai.kotlinlogging.KotlinLogging
+import io.ktor.http.ContentType.*
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -19,9 +21,13 @@ import net.folivo.trixnity.client.room.message.file
 import net.folivo.trixnity.client.room.message.image
 import net.folivo.trixnity.client.room.message.video
 import net.folivo.trixnity.core.model.RoomId
+import net.folivo.trixnity.utils.ByteArrayFlow
 import net.folivo.trixnity.utils.byteArrayFlowFromSource
+import net.folivo.trixnity.utils.toByteArray
+import net.folivo.trixnity.utils.toByteArrayFlow
 import okio.Buffer
 import org.koin.core.component.get
+
 
 private val log = KotlinLogging.logger { }
 
@@ -47,6 +53,7 @@ interface SendAttachmentViewModel {
     val isImage: Boolean?
     val isVideo: Boolean?
     val isAudio: Boolean?
+    val fileContent: StateFlow<ByteArrayFlow?>
 
     fun send()
     fun cancel()
@@ -71,6 +78,9 @@ class SendAttachmentViewModelImpl(
     override val isVideo = file.mimeType?.match("video/*")
     override val isAudio = file.mimeType?.match("audio/*")
 
+    private val _fileContent = MutableStateFlow<ByteArrayFlow?>(null)
+    override val fileContent: StateFlow<ByteArrayFlow?> = _fileContent.asStateFlow()
+
     private val backCallback = BackCallback {
         cancel()
     }
@@ -86,6 +96,18 @@ class SendAttachmentViewModelImpl(
                 _error.value = i18n.attachmentSizeMaxSizeError(messengerConfiguration.attachmentMaxSize)
             }
             _sendEnabled.value = _error.value == null
+            _fileContent.value = if (isImage == true && !checkFileSizeExceedsLimit(
+                    file.fileSize,
+                    messengerConfiguration.imageAttachmentMaxProcessingSize
+                )
+            ) get<ProcessImageUpload>().invoke(file.content.toByteArray(), file.mimeType ?: Image.PNG).toByteArrayFlow()
+            else if (isImage == true) {
+                log.debug{"Uploaded image ${file.fileName} couldn't be processed because it exceeds file size limits, it will be sent without processing"}
+                file.content
+            }
+            else {
+                file.content
+            }
         }
     }
 
@@ -94,7 +116,7 @@ class SendAttachmentViewModelImpl(
             _sendEnabled.value = false
             coroutineScope.launch {
                 matrixClient.room.sendMessage(selectedRoomId) {
-                    val byteArrayFlow = file.content
+                    val byteArrayFlow = fileContent.value?:file.content
                     when {
                         isImage ?: false -> {
                             log.debug { "send an image" }
@@ -167,6 +189,7 @@ class PreviewSendAttachmentViewModel() : SendAttachmentViewModel {
     override val isImage: Boolean = true
     override val isVideo: Boolean = false
     override val isAudio: Boolean = false
+    override val fileContent: StateFlow<ByteArrayFlow?> = MutableStateFlow(null)
 
     override fun send() {
     }